+++ src/Producers/Producer.php
@@ -39,6 +39,10 @@ class Producer
             $conf->set($key, $value);
         }
 
+        foreach ($this->config->getConfigCallbacks() as $method => $callback) {
+            $conf->{$method}($callback);
+        }
+
         return $conf;
     }
 
@@ -117,6 +121,8 @@ class Producer
      */
     private function flush(): mixed
     {
+        $sleepMilliseconds = config('kafka.flush_retry_sleep_in_ms', 100);
+
         return retry(10, function () {
             $result = $this->producer->flush(1000);
 
@@ -127,6 +133,6 @@ class Producer
             $message = rd_kafka_err2str($result);
 
             throw CouldNotPublishMessage::withMessage($message, $result);
-        });
+        }, $sleepMilliseconds);
     }
 }